Fixed db session

+++ database/models.py
@@ -1,6 +1,7 @@
+# У models.py
 from sqlalchemy.ext.asyncio import AsyncAttrs
 from sqlalchemy.orm import DeclarativeBase
-from sqlalchemy import Column, Integer, String, DateTime, Text, Boolean
+from sqlalchemy import Column, Integer, String, DateTime, Text, Boolean, BigInteger  # Додаємо BigInteger
 from datetime import datetime
 from config import settings
 
@@ -11,7 +12,7 @@ class User(Base):
     __tablename__ = "users"
     
     id = Column(Integer, primary_key=True)
-    telegram_id = Column(Integer, unique=True, nullable=False)
+    telegram_id = Column(BigInteger, unique=True, nullable=False)  # Змінюємо на BigInteger
     username = Column(String(50), nullable=True)
     first_name = Column(String(50), nullable=True)
     created_at = Column(DateTime, default=datetime.utcnow)
@@ -22,7 +23,7 @@ class QRCode(Base):
     __tablename__ = "qr_codes"
     
     id = Column(Integer, primary_key=True)
-    user_id = Column(Integer, nullable=False)
+    user_id = Column(BigInteger, nullable=False)  # Змінюємо на BigInteger
     text_content = Column(Text, nullable=False)
     format = Column(String(10), nullable=False)
     size = Column(Integer, nullable=False)
